qemu_monitor: Added  QEMU's "request-ebpf" support.
Added code for monitor and monitor_json.
The "request-ebpf" return's eBPF binary object encoded in base64.

QEMU provides eBPF that can be loaded and passed to it from Libvirt.
QEMU requires exact eBPF program/maps, so it can be retrieved using QAPI.
To load eBPF program - administrative capabilities are required,
so Libvirt may load it and pass it to the QEMU instance.
For now, there is only "RSS"(Receive Side Scaling) for virtio-net eBPF program and maps.

Signed-off-by: Andrew Melnychenko <andrew@daynix.com>